train/requirements.txt
asposestorage==1.0.2
core==1.0.1
dotdict==0.1
dv==1.0.12
einops==0.6.1
h5py==3.9.0
kornia==0.7.0
matplotlib==3.7.2
numba==0.57.1
numpy==1.23.5
opencv_contrib_python==4.8.0.76
pandas==2.1.0
pathlib2==2.3.7.post1
Pillow==9.4.0
Pillow==10.0.1
plotly==5.16.1
pytorch_lightning==1.9.0
pytorch_utils==0.5.5
pytz==2022.7
PyYAML==6.0
PyYAML==6.0.1
scikit_learn==1.3.0
scipy==1.10.1
skimage==0.0
thop==0.1.1.post2209072238
torch==2.0.0
torch_geometric==2.3.1
torchvision==0.15.0
tqdm==4.64.1
